+++ src/backend/base/langflow/schema/schema.py
@@ -1,14 +1,13 @@
-from collections.abc import AsyncIterator, Generator, Iterator
+from collections.abc import Generator
 from enum import Enum
 from typing import Literal
 
-from loguru import logger
 from pydantic import BaseModel
-from pydantic.v1 import BaseModel as BaseModelV1
 from typing_extensions import TypedDict
 
 from langflow.schema.data import Data
 from langflow.schema.message import Message
+from langflow.schema.serialize import recursive_serialize_or_str
 
 INPUT_FIELD_NAME = "input_value"
 
@@ -113,38 +112,3 @@ def build_output_logs(vertex, result) -> dict:
         outputs |= {name: OutputValue(message=message, type=_type).model_dump()}
 
     return outputs
-
-
-def recursive_serialize_or_str(obj):
-    try:
-        if isinstance(obj, str):
-            return obj
-        if isinstance(obj, dict):
-            return {k: recursive_serialize_or_str(v) for k, v in obj.items()}
-        if isinstance(obj, list):
-            return [recursive_serialize_or_str(v) for v in obj]
-        if isinstance(obj, BaseModel | BaseModelV1):
-            if hasattr(obj, "model_dump"):
-                obj_dict = obj.model_dump()
-            elif hasattr(obj, "dict"):
-                obj_dict = obj.dict()
-            return {k: recursive_serialize_or_str(v) for k, v in obj_dict.items()}
-
-        if isinstance(obj, AsyncIterator | Generator | Iterator):
-            # contain memory addresses
-            # without consuming the iterator
-            # return list(obj) consumes the iterator
-            # return f"{obj}" this generates '<generator object BaseChatModel.stream at 0x33e9ec770>'
-            # it is not useful
-            return "Unconsumed Stream"
-        if hasattr(obj, "dict"):
-            return {k: recursive_serialize_or_str(v) for k, v in obj.dict().items()}
-        if hasattr(obj, "model_dump"):
-            return {k: recursive_serialize_or_str(v) for k, v in obj.model_dump().items()}
-        if isinstance(obj, type) and issubclass(obj, BaseModel):
-            # This a type BaseModel and not an instance of it
-            return repr(obj)
-        return str(obj)
-    except Exception:  # noqa: BLE001
-        logger.debug(f"Cannot serialize object {obj}")
-        return str(obj)
